refactor: split glass module into action and drag submodules
Mirror the detached-glass structure: break the monolithic glass/module.js
into glass/action.js (minimize/restore/disabled-state) and glass/drag.js
(moved from the top-level draggable.js, since drag is a glass concern).
module.js is now a thin aggregator exposing enableGlassFeature(), and
binary-window collapses enableDrag()+enableActions() into that one call.
